Split tagfarm.utils off from tagfarm.main.
